% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{agents}
\alias{agents}
\title{Chemical agents}
\format{
Data frame of 329 rows whose names are agent identifiers, and 10
columns:
\describe{
\item{\code{name}}{\emph{Character}. Name of chemical agent.}
\item{\code{family}}{\emph{Factor}. Chemical family.}
\item{\code{molecular_weight}}{\emph{Numeric}. Molar mass, expressed in grams per
mole.}
\item{\code{size_fraction}}{\emph{Factor}. Particle size fraction considered.}
\item{\code{oel_source}}{\emph{Character}. Source of the exposure limit.}
\item{\code{oel}}{\emph{Numeric}. Occupational Exposure Limit according to the
source.}
\item{\code{oel_unit}}{\emph{Factor}. Unit in which the OEL is expressed.
\itemize{
\item F: fibres per cubic centimetre.
\item M: milligrams per cubic metre.
}}
\item{\code{casd_id_oel}}{\emph{Character}. Identifier of the CASD substance
considered for the corresponding exposure limit used in France.}
\item{\code{mixie_qc_substance}}{\emph{Character}. Name of the substance considered
for the corresponding toxicological classes from MiXie-QC.}
\item{\code{casd_id_mixie}}{\emph{Character}. Identifier of the CASD substance
considered for the corresponding toxicological classes from
MiXie-FR.}
}
}
\source{
Agent identifiers and names were created based on the original
substances they group together.
Chemical families were partly obtained from Bosson-Rieutort et al. (see
'References' section) and were assigned in a similar way for agents not
considered in this paper.

Molecular weights were obtained from an extract from 2007 of the PHYSPROP
database hosted by the Syracuse Research Corporation, from the OSHA
Occupational Chemical Database, and from the BOC Sciences website.
Sources for size fraction, OEL and units are given in column \code{oel_source}.
They were mainly obtained from the OSHA Occupational Chemical Database and
the GESTIS Substance Database.
CASD identifiers were provided by the \href{http://en.inrs.fr/}{INRS}.
MiXie-QC substance names were provided by the
\href{https://www.irsst.qc.ca/en/}{IRSST}.
}
\usage{
agents
}
\description{
Classification grouping substances from several occupational exposure
databases.
Associate chemical agents with properties, exposure limits and substances
from other classifications.
}
\details{
Different CASD substance identifiers may be associated with a single agent
for exposure limit and toxicity (i.e., \code{casd_id_oel} and \code{casd_id_mixie})
because chemicals agents may correspond to several CASD substances. Only one
limit is retained and the most relevant substance is considered for
toxicity.
}
\references{
Bosson-Rieutort D, Sarazin P, Bicout DJ, Ho V, Lavoué J (2020).
Occupational Co-exposures to Multiple Chemical Agents from Workplace
Measurements by the US Occupational Safety and Health Administration.
\emph{Annals of Work Exposures and Health}, Volume 64, Issue 4, May 2020,
Pages 402–415. \url{https://doi.org/10.1093/annweh/wxaa008}.
}
\seealso{
\code{\link{CASD_substances}}, \code{\link{mixie_qc}}, \code{\link{original_substances}},
\code{\link{substance_groups}}.
}
\keyword{datasets}
