% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CASD_substances}
\alias{CASD_substances}
\title{CASD substances}
\format{
Data frame of 284 rows whose names are CASD substance identifiers,
and 3 columns:
\describe{
\item{\code{oel_fr}}{\emph{Numeric}. Occupational Exposure Limit used in France
(8h-OEL).}
\item{\code{oel_fr_unit}}{\emph{Factor}. Unit in which the OEL is expressed.
\itemize{
\item F: fibres per cubic centimetre.
\item M: milligrams per cubic metre.
}}
\item{\code{mixie_fr_substance}}{\emph{Character}. Corresponding substance in
MiXie-FR.}
}
}
\source{
Data provided by the \href{http://en.inrs.fr/}{INRS}.
}
\usage{
CASD_substances
}
\description{
Substances from the CASD database.
Associate CASD substances with French occupational exposure limits and
substances from MiXie-FR.
}
\details{
It does not contain all substances from CASD.
}
\seealso{
\code{\link{mixie_fr}}, \code{\link{agents}}.
}
\keyword{datasets}
