% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{databases}
\alias{databases}
\title{Databases}
\format{
Character matrix of 4 rows whose names are database identifiers, and
3 columns:
\describe{
\item{\code{name}}{Name of the database.}
\item{\code{territory}}{Territory covered by the database.}
\item{\code{organization}}{Organization that hosts the database.}
}
}
\usage{
databases
}
\description{
Occupational exposure databases.
Associate databases with territories and organizations.
}
\details{
The USIS database combines the IMIS (Integrated Management Information
System) and the OIS (OSHA Information System) databases.
}
\seealso{
\code{\link{original_substances}}
}
\keyword{datasets}
