% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_database}
\alias{get_database}
\title{Get database identifiers}
\usage{
get_database(group_id = NULL)
}
\arguments{
\item{group_id}{Substance group identifiers. Any subset of \code{substance_groups}
names. If \code{NULL}, all groups are considered.}
}
\value{
Factor whose levels are \code{databases} rownames.
Identifier of the database corresponding to each given substance group.
}
\description{
Extract database identifiers of the given intra-database substance groups.
}
\examples{
## Search for databases of some substance groups then of all groups
get_database(names(substance_groups)[6:10])
get_database()

}
\author{
Gauthier Magnin
}
