% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_fr_substance}
\alias{get_fr_substance}
\title{Get MiXie-FR substances}
\usage{
get_fr_substance(agent_id = NULL)
}
\arguments{
\item{agent_id}{Agent identifiers. Any subset of \code{agents} rownames.
If \code{NULL}, all agents are considered.}
}
\value{
Name of the MiXie-FR substance associated with each given agent.
}
\description{
Extract MiXie-FR substances associated with the given agents.
}
\examples{
## Search for MiXie-FR substances of some agents then of all agents
get_fr_substance(rownames(agents)[6:10])
get_fr_substance()

}
\seealso{
\code{\link{get_fr_toxicity}}
}
\author{
Gauthier Magnin
}
