% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_qcfr_toxicity}
\alias{get_qcfr_toxicity}
\alias{get_qc_toxicity}
\alias{get_fr_toxicity}
\title{Get MiXie-QC toxicity or MiXie-FR toxicity}
\usage{
get_qc_toxicity(agent_id = NULL, as_vector = TRUE, id_only = TRUE)

get_fr_toxicity(agent_id = NULL, as_vector = TRUE, id_only = TRUE)
}
\arguments{
\item{agent_id}{Agent identifiers. Any subset of \code{agents} rownames.
If \code{NULL}, all agents are considered.}

\item{as_vector}{Ignored if more than one identifier is given.
\code{TRUE} or \code{FALSE} whether to return the result as a vector or as a named
list of a single vector.}

\item{id_only}{\code{TRUE} or \code{FALSE} whether to search for class identifiers
only or to search for class names too.}
}
\value{
Named list or vector (according to argument \code{as_vector}).
Toxicological classes associated with each given agent.
}
\description{
Extract MiXie-QC or MiXie-FR toxicological classes associated with the given
agents.
}
\examples{
## Search for some MiXie-QC toxicological classes
get_qc_toxicity(rownames(agents)[6:10])

## Search for toxicological classes of a single agent
## as a list or as a vector
get_qc_toxicity("0010", as_vector = FALSE)
get_qc_toxicity("0010", as_vector = TRUE)

## Search for toxicological classes, identifiers and names
get_qc_toxicity("0010", id_only = FALSE)

## Search for some MiXie-FR toxicological classes
get_fr_toxicity(rownames(agents)[6:10])

}
\seealso{
\code{\link{get_toxicity_names}}, \code{\link{get_fr_substance}}.
}
\author{
Gauthier Magnin
}
