% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_qcfr_toxicity_names}
\alias{get_qcfr_toxicity_names}
\alias{get_qc_toxicity_names}
\alias{get_fr_toxicity_names}
\title{Get toxicity names of MiXie-QC classes or of MiXie-FR classes}
\usage{
get_qc_toxicity_names(toxicity_id = NULL, french = FALSE)

get_fr_toxicity_names(toxicity_id = NULL, french = FALSE)
}
\arguments{
\item{toxicity_id}{Toxicological class identifiers.
For MiXie-QC classes, any subset of \code{toxicological_classes} rownames.
For MiXie-FR classes, any subset of column \code{mixie_fr_id} in
\code{toxicological_classes}.
If \code{NULL}, all classes are considered.}

\item{french}{\code{TRUE} or \code{FALSE} whether to extract French names or English
names.}
}
\value{
Name corresponding to each given toxicological class.
}
\description{
Extract names of toxicological classes.
}
\note{
These functions are not exported from the package.
}
\examples{
## Search for names of some MiXie-QC toxicological classes
get_qc_toxicity_names(rownames(toxicological_classes)[6:10])

## Search for names of some MiXie-FR toxicological classes
get_fr_toxicity_names(toxicological_classes[6:10, "mixie_fr_id"])

}
\seealso{
\code{\link{get_toxicity_names}}, \code{\link{get_qc_toxicity}}, \code{\link{get_fr_toxicity}}.
}
\author{
Gauthier Magnin
}
\keyword{internal}
