% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_toxicity_names}
\alias{get_toxicity_names}
\title{Get toxicity names}
\usage{
get_toxicity_names(toxicity_id, french = FALSE)
}
\arguments{
\item{toxicity_id}{Toxicological class identifiers. Any subset of the
rownames of \code{toxicological_classes} or of its column \code{mixie_fr_id}.}

\item{french}{\code{TRUE} or \code{FALSE} whether to extract French names or English
names.}
}
\value{
Name corresponding to each given toxicological class.
}
\description{
Extract names of toxicological classes.
}
\examples{
## Search for names of some MiXie-QC or MiXie-FR toxiciological classes
get_toxicity_names(rownames(toxicological_classes)[6:10])
get_toxicity_names(toxicological_classes[6:10, "mixie_fr_id"])

}
\seealso{
\code{\link{get_qc_toxicity}}, \code{\link{get_fr_toxicity}}.
}
\author{
Gauthier Magnin
}
