% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{join_agent_data}
\alias{join_agent_data}
\title{Join agent data}
\usage{
join_agent_data(tox_names = TRUE)
}
\arguments{
\item{tox_names}{\code{TRUE} or \code{FALSE} whether to also include toxicological
class names or to only include class identifiers.}
}
\value{
A single data frame combining most datasets from the package.
Its row names are agent identifiers.
}
\description{
Join agent data into a single data frame. Consider the main dataset (the one
named \code{agents}) and combine it with the corresponding values from other ones
(\code{CASD_substances}, \code{mixie_qc}, \code{mixie_fr} and \code{toxicological_classes}).
}
\details{
Regarding the dataset \code{agents}, additional columns are the following.
\describe{
\item{\code{oel_fr}}{Occupational Exposure Limit used in France corresponding to
substance identifiers from \code{agents$casd_id_oel}.}
\item{\code{oel_fr_unit}}{Units in which the OELs from \code{oel_fr} are expressed.
\itemize{
\item F: fibres per cubic centimetre.
\item M: milligrams per cubic metre.
}}
\item{\code{toxicity_qc}}{Toxiciological classes from MiXie-QC corresponding to
substances from \code{agents$mixie_qc_substance}.}
\item{\code{mixie_fr_substance}}{Substance in MiXie-FR corresponding to
substance identifiers from \code{agents$casd_id_mixie}.}
\item{\code{toxicity_fr}}{Toxicological classes from MiXie-FR corresponding to
substances from \code{mixie_fr_substance}.}
}

Data from column \code{name_fr} of dataset \code{toxicological_classes} are not
included in the resulting data frame.
Data from datasets \code{original_substances}, \code{databases} and \code{substance_groups}
are not included either.
}
\examples{
## Create the data frame of agent data
agent_data <- join_agent_data()

}
\seealso{
\code{\link{join_all_data}}, \code{\link{join_grouping_data}}.
}
\author{
Gauthier Magnin
}
