% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{join_all_data}
\alias{join_all_data}
\title{Join data}
\usage{
join_all_data()
}
\value{
List of two data frames, containing agent data and grouping data.
}
\description{
Join (almost) all datasets from the package into two data frames.
One contains agent data. The other one contains grouping data about
original substances in occupational exposure databases.
}
\details{
Agent data (rownames are agent identifiers):
\describe{
\item{\code{name}}{Names of the chemical agents.}
\item{\code{family}}{Chemical families.}
\item{\code{molecular_weight}}{Molar masses, expressed in grams per mole.}
\item{\code{size_fraction}}{Particle size fractions considered.}
\item{\code{oel_source}}{Sources of the exposure limits.}
\item{\code{oel}}{Occupational Exposure Limits according to the sources
(8h-OELs).}
\item{\code{oel_unit}}{Units in which the OELs are expressed.
\itemize{
\item F: fibres per cubic centimetre.
\item M: milligrams per cubic metre.
}}
\item{\code{casd_id_oel}}{Identifiers of the CASD substances considered for the
corresponding exposure limits used in France.}
\item{\code{oel_fr}}{Occupational Exposure Limits used in France (8h-OELs).}
\item{\code{oel_fr_unit}}{Units in which the OELs used in France are expressed.
\itemize{
\item F: fibres per cubic centimetre.
\item M: milligrams per cubic metre.
}}
\item{\code{mixie_qc_substance}}{Names of the substances considered for the
corresponding toxicological classes from MiXie-QC.}
\item{\code{toxicity_qc}}{Toxiciological classes corresponding to the substances
of MiXie-QC considered.}
\item{\code{casd_id_mixie}}{Identifiers of the CASD substances considered for
the corresponding toxicological classes from MiXie-FR.}
\item{\code{mixie_fr_substance}}{Corresponding substances in MiXie-FR.}
\item{\code{toxicity_fr}}{Toxicological classes corresponding to the substances
of MiXie-FR considered.}
}

Grouping data:
\describe{
\item{\code{original_id}}{Identifiers of the original substances.}
\item{\code{database_id}}{Identifiers of the databases from which the original
substances come.}
\item{\code{group_id}}{Group identifiers of substances from the same database
(i.e., intra-database groups).}
\item{\code{agent_id}}{Main chemical agent identifiers (i.e., the
inter-database groups) corresponding to the groups of original
substances (i.e., the intra-database groups).}
\item{\code{agent_name}}{Names of the chemical agents corresponding to the main
agent identifiers.}
}

Data not included in the resulting data frames are data from the dataset
\code{databases} and from the column \code{name_fr} of the dataset
\code{toxicological_classes}.
}
\examples{
## Create the list of two data frames and look at each one
data_list <- join_all_data()
View(data_list$agent)
View(data_list$grouping)

}
\seealso{
\code{\link{join_agent_data}}, \code{\link{join_grouping_data}}.
}
\author{
Gauthier Magnin
}
