% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{join_grouping_data}
\alias{join_grouping_data}
\title{Join grouping data}
\usage{
join_grouping_data()
}
\value{
A single data frame combining the datasets from the package that are
about groupings.
}
\description{
Join grouping data into a single data frame. Consider the dataset
\code{original_substances} and combine it with the corresponding values from
the datasets \code{substance_groups} and \code{agents}.
}
\details{
Regarding the dataset \code{\link{original_substances}}, additional columns are the
following.
\describe{
\item{\code{original_id}}{Identifiers of the original substances (row names from
\code{original_substances}).}
\item{\code{agent_id}}{Identifiers of the chemical agents (i.e., the
inter-database groups) corresponding to
\code{original_substances$group_id} (i.e., the intra-database groups).}
\item{\code{agent_name}}{Names of the chemical agents corresponding to
\code{agent_id}.}
}

Row names are only row numbers and do not represent any specific data.
}
\examples{
## Create the data frame of grouping data
grouping_data <- join_grouping_data()

}
\seealso{
\code{\link{join_all_data}}, \code{\link{join_agent_data}}.
}
\author{
Gauthier Magnin
}
