% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mixie_qc}
\alias{mixie_qc}
\title{MiXie Quebec}
\format{
Character matrix of 952 rows and 2 columns:
\describe{
\item{\code{substance}}{Substance name.}
\item{\code{toxicity_id}}{Identifier of a toxicological class.}
}
}
\source{
Data provided by the \href{https://www.irsst.qc.ca/en/}{IRSST}.
}
\usage{
mixie_qc
}
\description{
Substances and toxicological classes from MiXie-QC.
Associate substances with toxicological classes.
}
\details{
It does not contain all substances from MiXie-QC.
}
\seealso{
\code{\link{mixie_fr}}, \code{\link{toxicological_classes}}.
}
\keyword{datasets}
