% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{original_substances}
\alias{original_substances}
\title{Original substances}
\format{
Data frame of 1 065 rows whose names are original substance
identifiers, and 2 columns:
\describe{
\item{\code{database_id}}{\emph{Factor}. Identifier of the database from which the
substance comes.}
\item{\code{group_id}}{\emph{Character}. Group identifier of substances from the
same database.}
}
}
\source{
Substance identifiers were provided
by the \href{https://www.irsst.qc.ca/en/}{IRSST} for the LIMS database and
by the \href{http://en.inrs.fr/}{INRS} for the CASD database.
Substance identifiers for the USIS database are part of the results of
another project: see dataset \code{USIS_substances} from package
\href{https://github.com/UofMontreal-Multiexpo/uom.usis}{uom.usis}.
Substance identifiers for the CWED database were created.

Group identifiers were created.
}
\usage{
original_substances
}
\description{
Substances from several occupational exposure databases.
Associate substance identifiers with their original databases and with
identifiers of groups within those databases.
}
\details{
It does not contain all substances from each database.
}
\seealso{
\code{\link{databases}}, \code{\link{substance_groups}}.
}
\keyword{datasets}
