% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{substance_groups}
\alias{substance_groups}
\title{Substance groups}
\format{
Named character vector of 854 elements.
}
\source{
Group and agent identifiers were created.
}
\usage{
substance_groups
}
\description{
Substance groups from several databases.
Associate identifiers of intra-database substance groups with main chemical
agent identifiers, which can be considered as inter-database substance
groups.
}
\seealso{
\code{\link{agents}}, \code{\link{original_substances}}.
}
\keyword{datasets}
