% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{toxicological_classes}
\alias{toxicological_classes}
\title{Toxicological classes}
\format{
Character matrix of 24 rows whose names are toxicological class
identifiers from MiXie-QC, and 3 columns:
\describe{
\item{\code{mixie_fr_id}}{Identifier for the MiXie-FR toxicological class.}
\item{\code{name}}{Name of the toxicological class.}
\item{\code{name_fr}}{French name of the toxicological class.}
}
}
\source{
Data provided by the \href{https://www.irsst.qc.ca/en/}{IRSST} and the
\href{http://en.inrs.fr/}{INRS}, except identifiers of French toxicological
classes.

Identifiers of French toxicological classes were created for all provided
classes from MiXie-FR.
}
\usage{
toxicological_classes
}
\description{
Toxicological classes from MiXie-QC and MiXie-FR.
Associate class identifiers with English and French class names.
}
\seealso{
\code{\link{mixie_fr}}, \code{\link{mixie_qc}}.
}
\keyword{datasets}
