% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uom.agents-package.R
\docType{package}
\name{uom.agents-package}
\alias{uom.agents}
\alias{uom.agents-package}
\title{uom.agents: Chemical Agents for Occupational Exposure Databases}
\description{
Data about chemical agents relating to occupational exposure databases: classifications, exposure limit values, toxicological classes and other properties. Raw data and CSV format of final datasets are available in this \href{https://www.dropbox.com/scl/fo/0k1d84kf005qc6n7xgimh/AEbgRasCpHqdsSXvRCzV9e4?rlkey=8wl7i5yu8zfqgbrpmuybt7h5j&st=8x3td6md&dl=0}{Dropbox folder}.
}
\details{
For an organized list of datasets and functions, use
\code{vignette("list_of_help_pages", "uom.agents")}

For an alphabetical list and for the package description, use
\code{library(help = "uom.agents")}

For package news information, use \code{news(package = "uom.agents")}

For information on how to cite the package, use \code{citation("uom.agents")}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/UofMontreal-Multiexpo/uom.agents}
  \item Report bugs at \url{https://github.com/UofMontreal-Multiexpo/uom.agents/issues}
}

}
\author{
\strong{Maintainer}: Jérôme Lavoué \email{jerome.lavoue@umontreal.ca}

Authors:
\itemize{
  \item Gauthier Magnin [contractor]
}

Other contributors:
\itemize{
  \item CHUM - Centre Hospitalier de l'Université de Montréal [copyright holder, funder]
  \item IRSST - Institut de Recherche Robert-Sauvé en Santé et en Sécurité du Travail [copyright holder, funder, data contributor]
  \item INRS - Institut National de Recherche et de Sécurité pour la prévention des accidents du travail et des maladies professionnelles [data contributor]
}

}
